Got it up and running!
You can make docs and edit them now. CRU, no D yet. But this is just a
guy playing around so things will move a bit slowly. I've got a long way
to go to get this up to the system I dream it to be, but I am pleased
with what I've got so far.