Directory viewer with links works.
I'm not sure if it would work with sub directories, it can't make new
documents and it can't make new folders, but it does show the existing
documents, which is good enough for now. I really want to get back to
the editor so I'm happy with something that does the proper querying and
calling that good enough.

This isn't technically a single page app. While the same root javascript
file gets called each time, variables passed in change slightly
depending on your URL so I can get navigation done that way. It's not
the way you are supposed to do React apps but it is quick and easy. I'll
come back to get this a single page web app later.

I have created an editor component, my hope is that it will have many
sub-editors to choose from. For now I'll be happy just to get back to
the basic functionality of the automatic saving.

Oh yea, and eventually I'll get back to doing patches to update
documents rather than sending the entire document. I'm not currently
even sending anything, so, baby steps. XD