Time limiting mostly works.
I've setup a system that will choose to skip a post if it has been less
than two seconds. I'll turn that value into a variable later, but for
now it works as is. There is one problem though. Any edits you make in
the two seconds after a save won't be saved. I'll fix that later.

For now, I think I've finally cleared the way to work on posting diffs.