Got diff in, starting to use.
I spent the rest of my evening just getting diff to import. I only got
it to import using a require. I'm not sure how to change it to an import
but I can cross that bridge later. For now I'll just see about getting
this to be actually pushed.