First step in save by diffs done.
I can now actually do the save by diff. It's interesting to wonder if it
will actually save much data in the long run, but I like the idea of
using it for undos as well as saving data so there is more than one
feature being tackled with this feature.

Now though it is critical that the chain not be broken. Every diff has
to be linked to the appropriate parent or the document will not be able
to do/undo propperly. Which means I need to implment a stack of diffs in
the editor, which I need for undo anyway, so that when there is a hiccup
in transfer I can resend. My plan is to reject a diff with an unknown
parent and send the latest diff that the server has back. Then the UI
can step through its stack of diffs until it find the the diff in the
stack with the returned diff as a parent and then send the whole stack
from that point to the present.

I realized that such will be necessary when I was writing a document and
wandered out of wifi range. Since my personal server is running a older
vision of Bento which saves the entire document, that wasn't an issue,
but if I had been saving by diffs it would have caused a real problem.