Save by diff recoverable.
Now if you loose connection for some reason Bento will attempt to
recover by pushing the entire diff stack over. It's not elegant and has
a couple of problems, but it is basic functionality and keeps the diff
chain intact.

Now for a few creature comforts, Lets update the editor.