Moved update to interval
I have moved the update function back to setInterval so that it can run
every two seconds regardless of whether the user has typed anything or
not. This means that you no longer have to type, wait to seconds, and
then type one more character to get all of your content in.

My next trick is going to be synchronizing down from the server. Right
now I am primarily using this as a clipboard between the 4 logical
computers I use at work. When I paste something in said clip board I
need to refresh the page to get it to show up. Lets fix that.