Had an idea while I was falling asleep.
It isn't the most elegant way to do this, but it works. The problem is
that it doesn't do any sort of conflict resolution, rather, it will just
drop any changes you have made locally. But I don't expect a lot of
cases where people are working on the same document at the same time, or
worse offline for a while editing. But it works for now which is all I
need.