Syncing was a little hecktick.
The syncing was a little much. If user was editing at the wrong time it
would result in the editor thinking it is out of sync with the server
since it would be on a diff that is different from the server. I
resolved this by having it check if the from the server is in the stack
of diffs in the editor. This basically resolved the issue. There is
still a problem of the user's cursor jumping to the top when this
happens. Not ideal but a tolerable bug for now.