Moved View out of Directory.
I have broken Directory into two parts so far. Itself and then a view
manager that will be managing which view of the items in a folder the
user wants. At least one more component will be coming in the form of a
menu bar, but I'll cross that bridge when I come to it.