"Slide in menu" working
It doesn't slide. I'm not particularly great at CSS annimations yet. And
that's polish that isn't necessary for it to look good enough. I make no
claims of being a designer. But the slide in menu does work and the new
button creates a new document so I am back up to functional. There is
still a bit of work to do in many was to make it nice, but I'm pretty
happy with what I have accomplished this evening.