Partial dark theme.
I realized the for the most part editors will need to be light themed.
This is just the nature of the types of editors I'm planning on working
with. Word and Excel are write and trying to do something different is
going to require me to code up my own solutions for those editors which
I'd rather avoid. So I'm trying to give the currently open display the
ability to control the color. What I've come up with so far isn't great
but it does achieve the goal of dark theme file manager light them
editor.