Trying to add context menu.
Strictly speaking not exactly a context menu. I'm trying to add a menu
that appears when you click the vertical ellipsis. This menu will have
the move, rename, and delete methods. I'm not sure how to go about
creating said menu at this point though. I've got the vertical ellipsis
which is a good start though.

Note: I'm hoping to turn this into a progressive web app, and from there
a Trusted Web Action so I can compile it into an apk. Because I'm using
one UI for desktop and mobile (as I am not a web designer) I'm
intentionally avoiding actions that aren't available on both.