Menu exists, rename is clickable.
I'm not sure why the rest of it isn't clickable, but rename is. I also
couldn't figure out how to have it show up on the right so I moved the
menu button to the left. There is still a fair bit of work to do here
but I wanted to save here with something working.